+++ src/Speed/GameCube/Src/Logitech/Constant.cpp
@@ -18,7 +18,7 @@ static inline unsigned long &ConstantGetEffectID(Force *self, int channel, int f
 
 Constant::Constant() : Force() {}
 
-int Constant::DownloadForce(int channel, int forceNumber, unsigned long & handle, unsigned long duration, unsigned long startDelay, short magnitude, unsigned short direction, unsigned long attackTime, unsigned long fadeTime, unsigned char attackLevel, unsigned char fadeLevel) {
+int Constant::DownloadForce(long channel, long forceNumber, unsigned long & handle, unsigned long duration, unsigned long startDelay, short magnitude, unsigned short direction, unsigned long attackTime, unsigned long fadeTime, unsigned char attackLevel, unsigned char fadeLevel) {
     LGForceEffect force;
     int ret;
 
@@ -51,7 +51,7 @@ int Constant::DownloadForce(int channel, int forceNumber, unsigned long & handle
     return ret;
 }
 
-int Constant::UpdateForce(int channel, int forceNumber, unsigned long duration, unsigned long startDelay, short magnitude, unsigned short direction, unsigned long attackTime, unsigned long fadeTime, unsigned char attackLevel, unsigned char fadeLevel) {
+int Constant::UpdateForce(long channel, long forceNumber, unsigned long duration, unsigned long startDelay, short magnitude, unsigned short direction, unsigned long attackTime, unsigned long fadeTime, unsigned char attackLevel, unsigned char fadeLevel) {
     LGForceEffect force;
     int ret;
