+++ src/Speed/GameCube/Src/Logitech/Wheels.cpp
@@ -61,17 +61,17 @@ short Wheels::ReadAll() {
     return wheelUnplugged;
 }
 
-bool Wheels::ButtonIsPressed(int channel, unsigned long buttonMask) {
+bool Wheels::ButtonIsPressed(long channel, unsigned long buttonMask) {
     const LGPosition *position = reinterpret_cast<const LGPosition *>(this);
     return (position[channel].button & buttonMask) != 0;
 }
 
-bool Wheels::IsConnected(int channel) {
+bool Wheels::IsConnected(long channel) {
     const LGPosition *position = reinterpret_cast<const LGPosition *>(this);
     return position[channel].err != 0;
 }
 
-bool Wheels::PedalsConnected(int channel) {
+bool Wheels::PedalsConnected(long channel) {
     const LGPosition *position = reinterpret_cast<const LGPosition *>(this);
     return (position[channel].misc >> 3) & 1;
 }
